date format updated

+++ server/migration.sql
@@ -9,7 +9,7 @@ CREATE TABLE rentals (
   date DATE,
   group_size INTEGER,
   type TEXT,
-  rating INTEGER,
+  rating DECIMAL (2,1),
   review INTEGER,
   description TEXT,
   image VARCHAR 
@@ -23,7 +23,7 @@ CREATE TABLE my_rentals (
   date DATE,
   group_size INTEGER,
   type TEXT,
-  rating INTEGER,
+  rating DECIMAL (2,1),
   review INTEGER,
   description TEXT,
   image VARCHAR
